+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to
 - ✨(frontend) create skeleton component for DocEditor #1491
 - ✨(frontend) add an EmojiPicker in the document tree and title #1381
 - ✨(frontend) ajustable left panel #1456
+- ✨(frontend) improve mobile UX by showing subdocs count #1540
 
 ### Changed